* move CreateStreamFromName() to own header file to allow use in integration tests
